Resolved Client Block and started whisper recognition
The client was blocking either on searching for a new message from the
server, or searching for a new message form the client.  I split it into
two processes so each could take care of their projects individually.
The child is not receiving message from the server atm.  I tried
shifting clientfd into shared memory to resolve the issue, but it didn't
work.

On the server I added a method "checkForWhisper" which would look at
each string to be broadcasted and determine if it was a whisper,
returning the address of the recipient if it was a whisper, 0 if it was
not.  Currently strtok is destroying my message, so I started trying to
copy message to a new char* to make sure the original message stays
intact.