whisper recognition
when a client sends a string of format "@x:message" where x is the id of
the recipient, the server extracts x to as the address to send the
message to.  Currently the server identifies x as the clientfd, not the
bucket of the clients array.

the client still is not receiving messages from the server.