added signals and fixed client recv error
added a while loop around the process in client.c so that it continued
checking for new messages after the first one.

Added a signal to server.c that is called when a new client connects.
The server then broadcasts that a new client has connected to all
clients. 2 known issues:
1) The child processes of server.c do not know about the connected
clients (I thought they should, but the fix eludes me)
2) The broadcast when a new client connects is a simple notification
that does not inform the clients who the client that connected is.
Therefore, they have no way to contact him directly.